Rename packages to br.com.amz.ktorplate
