+++ application/src/main/resources/application.conf
@@ -4,7 +4,7 @@ ktor {
     }
 
     application {
-        modules = [ br.com.stonks.poc.ktor.web.server.ModuleKt.module ]
+        modules = [ br.com.amz.ktorplate.web.server.ModuleKt.module ]
     }
 }
