+++ entity/src/main/kotlin/br/com/amz/ktorplate/exception/HttpExceptions.kt
@@ -1,0 +1,7 @@
+package br.com.amz.ktorplate.exception
+
+class UnauthorizedException(override val message: String = ""): Throwable(message)
+
+class PreConditionFailedException(override val message: String = ""): Throwable(message)
+
+class NotFoundException(override val message: String = ""): Throwable(message)
