+++ entity/src/main/kotlin/br/com/amz/ktorplate/user/User.kt
@@ -1,0 +1,10 @@
+package br.com.amz.ktorplate.user
+
+data class User(
+    val id: Long? = null,
+    val name: String? = null,
+    val email: String,
+    val phone: String? = null,
+    val password: String,
+    val active: Boolean = true
+)
