--- entity/src/main/kotlin/br/com/stonks/poc/ktor/exception/HttpExceptions.kt
@@ -1,7 +1,0 @@
-package br.com.stonks.poc.ktor.exception
-
-class UnauthorizedException(override val message: String = ""): Throwable(message)
-
-class PreConditionFailedException(override val message: String = ""): Throwable(message)
-
-class NotFoundException(override val message: String = ""): Throwable(message)
