+++ usecase/src/main/kotlin/br/com/amz/ktorplate/usecases/service/UserService.kt
@@ -1,0 +1,9 @@
+package br.com.amz.ktorplate.usecases.service
+
+import br.com.amz.ktorplate.usecases.adapter.UserAdapter
+
+class UserService(private val userAdapter: UserAdapter) {
+    suspend fun findAll() = userAdapter.findAll()
+
+    suspend fun findByEmail(email: String) = userAdapter.findByEmail(email)
+}
