--- usecase/src/main/kotlin/br/com/stonks/poc/ktor/usecases/adapter/UserAdapter.kt
@@ -1,11 +1,0 @@
-package br.com.stonks.poc.ktor.usecases.adapter
-
-import br.com.stonks.poc.ktor.user.User
-
-interface UserAdapter {
-    suspend fun findAll(): List<User>
-
-    suspend fun findByEmail(email: String): User?
-
-    suspend fun save(user: User): Long
-}
