+++ web/src/main/kotlin/br/com/amz/ktorplate/web/dto/LoginRequest.kt
@@ -1,0 +1,10 @@
+package br.com.amz.ktorplate.web.dto
+
+data class SignUpRequest(
+    val name: String,
+    val email: String,
+    val phone: String,
+    val password: String
+)
+
+data class LoginRequest(val email: String, val password: String)
