+++ web/src/main/kotlin/br/com/amz/ktorplate/web/server/ServerBoot.kt
@@ -1,0 +1,10 @@
+package br.com.amz.ktorplate.web.server
+
+import io.ktor.server.netty.EngineMain
+
+object ServerBoot {
+
+    fun boot(args: Array<String>){
+        EngineMain.main(args)
+    }
+}
