--- web/src/main/kotlin/br/com/stonks/poc/ktor/web/config/KoinWebModule.kt
@@ -1,10 +1,0 @@
-package br.com.stonks.poc.ktor.web.config
-
-import br.com.stonks.poc.ktor.web.controller.LoginController
-import br.com.stonks.poc.ktor.web.controller.UserController
-import org.koin.dsl.module
-
-val controllerModule = module {
-    single { LoginController(get()) }
-    single { UserController(get()) }
-}
