--- web/src/main/kotlin/br/com/stonks/poc/ktor/web/server/ServerBoot.kt
@@ -1,10 +1,0 @@
-package br.com.stonks.poc.ktor.web.server
-
-import io.ktor.server.netty.EngineMain
-
-object ServerBoot {
-
-    fun boot(args: Array<String>){
-        EngineMain.main(args)
-    }
-}
