+++ web/src/test/kotlin/br/com/amz/ktorplate/ApplicationTest.kt
@@ -1,0 +1,28 @@
+package br.com.amz.ktorplate
+
+import br.com.amz.ktorplate.web.server.module
+import io.ktor.http.HttpMethod
+import io.ktor.http.HttpStatusCode
+import io.ktor.locations.KtorExperimentalLocationsAPI
+import io.ktor.server.testing.handleRequest
+import io.ktor.server.testing.withTestApplication
+import io.ktor.util.KtorExperimentalAPI
+import org.junit.Ignore
+import org.junit.Test
+import kotlin.test.assertEquals
+
+@KtorExperimentalAPI
+@KtorExperimentalLocationsAPI
+class ApplicationTest {
+
+    @Ignore
+    @Test
+    fun testRoot() {
+        withTestApplication({ module() }) {
+            handleRequest(HttpMethod.Get, "/health").apply {
+                assertEquals(HttpStatusCode.OK, response.status())
+                assertEquals("{\"status\":\"OK\"}", response.content)
+            }
+        }
+    }
+}
