--- web/src/test/kotlin/br/com/stonks/poc/ktor/ApplicationTest.kt
@@ -1,28 +1,0 @@
-package br.com.stonks.poc.ktor
-
-import br.com.stonks.poc.ktor.web.server.module
-import io.ktor.http.HttpMethod
-import io.ktor.http.HttpStatusCode
-import io.ktor.locations.KtorExperimentalLocationsAPI
-import io.ktor.server.testing.handleRequest
-import io.ktor.server.testing.withTestApplication
-import io.ktor.util.KtorExperimentalAPI
-import org.junit.Ignore
-import org.junit.Test
-import kotlin.test.assertEquals
-
-@KtorExperimentalAPI
-@KtorExperimentalLocationsAPI
-class ApplicationTest {
-
-    @Ignore
-    @Test
-    fun testRoot() {
-        withTestApplication({ module() }) {
-            handleRequest(HttpMethod.Get, "/health").apply {
-                assertEquals(HttpStatusCode.OK, response.status())
-                assertEquals("{\"status\":\"OK\"}", response.content)
-            }
-        }
-    }
-}
